Split GraphiQLMenuPage into class and extension
